+++ internal/models/group.go
@@ -1,7 +1,7 @@
 package models
 
 type GroupedBalance struct {
-	Id       int
+	Id       int64
 	Balances *map[string]*Balance
 }
 
@@ -12,7 +12,7 @@ func MakeNewGroupBalance(balances *map[string]*Balance) *GroupedBalance {
 	}
 }
 
-func (group *GroupedBalance) SetId(id int) {
+func (group *GroupedBalance) SetId(id int64) {
 	group.Id = id
 }
 
@@ -21,15 +21,15 @@ func (group *GroupedBalance) GetBalance(userId string) bool {
 	return ok
 }
 
-func (group *GroupedBalance) SetBalance(userId string, startTime int) {
+func (group *GroupedBalance) SetBalance(userId string, startTime int64) {
 	(*group.Balances)[userId] = MakeBalance(startTime)
 }
 
-func (group *GroupedBalance) UpdateBalance(userId string, value float64, time int) {
+func (group *GroupedBalance) UpdateBalance(userId string, value float64, time int64) {
 	(*group.Balances)[userId].Update(value, time)
 }
 
-func (group *GroupedBalance) UpdateLastAll(lastTime int) {
+func (group *GroupedBalance) UpdateLastAll(lastTime int64) {
 	for _, balance := range *group.Balances {
 		balance.UpdateLast(lastTime)
 	}
