+++ internal/models/transaction.go
@@ -1,11 +1,11 @@
 package models
 
 type Transaction struct {
-	Timestamp int
+	Timestamp int64
 	Price     float32
 }
 
-func MakeTransaction(time int, price float32) *Transaction {
+func MakeTransaction(time int64, price float32) *Transaction {
 	return &Transaction{
 		Timestamp: time,
 		Price:     price,
